chore: Recover `@rocket.chat/fuselage-ui-kit` storybook
