+++ packages/fuselage-ui-kit/.storybook/main.ts
@@ -1,12 +1,19 @@
-import { type StorybookConfig } from '@storybook/core-common';
+import type { StorybookConfig } from '@storybook/core-common';
 
-const config: StorybookConfig = {
+export default {
   stories: ['../src/**/*.stories.tsx', '../src/**/stories.tsx'],
-  addons: ['@storybook/addon-essentials', 'storybook-dark-mode'],
+  addons: [
+    '@storybook/addon-essentials',
+    'storybook-dark-mode',
+    {
+      name: '@newhighsco/storybook-addon-transpile-modules',
+      options: {
+        transpileModules: ['date-fns', 'typia', 'react-i18next'],
+      },
+    },
+  ],
   features: {
     postcss: false,
   },
   framework: '@storybook/react',
-};
-
-module.exports = config;
+} satisfies StorybookConfig;
