+++ packages/fuselage-ui-kit/.storybook/preview.tsx
@@ -1,11 +1,13 @@
 import { PaletteStyleTag } from '@rocket.chat/fuselage';
+import { DocsContainer as BaseContainer } from '@storybook/addon-docs';
 import { type Parameters } from '@storybook/addons';
 import { type DecoratorFn } from '@storybook/react';
 import { themes } from '@storybook/theming';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import type { ComponentPropsWithoutRef } from 'react';
 import { useDarkMode } from 'storybook-dark-mode';
 
 import manifest from '../package.json';
-import { DocsContainer } from './DocsContainer';
 import { surface } from './helpers';
 import logo from './logo.svg';
 
@@ -13,6 +15,42 @@ import '@rocket.chat/fuselage/dist/fuselage.css';
 import '@rocket.chat/icons/dist/rocketchat.css';
 import '@rocket.chat/fuselage-polyfills';
 import 'normalize.css/normalize.css';
+
+type DocsContainerProps = ComponentPropsWithoutRef<typeof BaseContainer>;
+
+const DocsContainer = (props: DocsContainerProps) => {
+  const dark = useDarkMode();
+
+  const { context } = props;
+
+  return (
+    <BaseContainer
+      {...props}
+      context={{
+        ...context,
+        storyById: (id) => {
+          const storyContext = context.storyById(id);
+          return {
+            ...storyContext,
+            parameters: {
+              ...storyContext?.parameters,
+              docs: {
+                ...storyContext?.parameters?.docs,
+                theme: dark
+                  ? {
+                      ...themes.dark,
+                      appContentBg: surface.main,
+                      barBg: surface.main,
+                    }
+                  : themes.light,
+              },
+            },
+          };
+        },
+      }}
+    />
+  );
+};
 
 export const parameters: Parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
@@ -49,6 +87,8 @@ export const parameters: Parameters = {
   },
 };
 
+const queryClient = new QueryClient();
+
 export const decorators: DecoratorFn[] = [
   (fn) => {
     const dark = useDarkMode();
@@ -60,4 +100,5 @@ export const decorators: DecoratorFn[] = [
       </>
     );
   },
+  (fn) => <QueryClientProvider client={queryClient} children={fn()} />,
 ];
