+++ packages/fuselage-ui-kit/package.json
@@ -37,9 +37,8 @@
     "testunit": "jest",
     "typecheck": "tsc --noEmit",
     "docs": "cross-env NODE_ENV=production build-storybook -o ../../static/fuselage-ui-kit",
-    "storybook": "start-storybook -p 6006 --no-version-updates",
-    "build-storybook": "cross-env NODE_ENV=production build-storybook",
-    "bump-next": "bump-next"
+    "storybook": "cross-env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' start-storybook -p 6006 --no-version-updates",
+    "build-storybook": "cross-env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' NODE_ENV=production build-storybook"
   },
   "peerDependencies": {
     "@rocket.chat/apps-engine": "*",
@@ -63,6 +62,7 @@
     "@babel/preset-env": "~7.22.20",
     "@babel/preset-react": "~7.22.15",
     "@babel/preset-typescript": "~7.22.15",
+    "@newhighsco/storybook-addon-transpile-modules": "~1.0.0",
     "@rocket.chat/apps-engine": "workspace:^",
     "@rocket.chat/core-typings": "workspace:^",
     "@rocket.chat/eslint-config": "workspace:^",
@@ -80,10 +80,7 @@
     "@rocket.chat/ui-video-conf": "workspace:^",
     "@storybook/addon-essentials": "~6.5.16",
     "@storybook/addons": "~6.5.16",
-    "@storybook/builder-webpack5": "~6.5.16",
-    "@storybook/manager-webpack5": "~6.5.16",
     "@storybook/react": "~6.5.16",
-    "@storybook/source-loader": "~6.5.16",
     "@storybook/theming": "~6.5.16",
     "@tanstack/react-query": "^4.16.1",
     "@testing-library/react": "~16.0.1",
@@ -102,15 +99,13 @@
     "react-dom": "^17.0.2",
     "react-i18next": "~13.2.2",
     "rimraf": "^3.0.2",
-    "storybook-dark-mode": "~3.0.3",
-    "tslib": "^2.5.3",
+    "storybook-dark-mode": "~2.1.1",
     "typescript": "~5.5.4"
   },
   "dependencies": {
     "@rocket.chat/core-typings": "workspace:^",
     "@rocket.chat/gazzodown": "workspace:^",
-    "@rocket.chat/ui-kit": "workspace:~",
-    "tslib": "^2.5.3"
+    "@rocket.chat/ui-kit": "workspace:~"
   },
   "volta": {
     "extends": "../../package.json"
